修改 新增/修改banner 可视规格bug

+++ src/main/java/io/github/talelin/latticy/model/SpuDO.java
@@ -37,7 +37,7 @@ public class SpuDO extends BaseModel {
 
     private Long rootCategoryId;
 
-    private Long sketchSpecId;
+    private Integer sketchSpecId;
 
     private String subtitle;
